프론트엔드 년도, 달 네비게이션 추가
- 기존에 네비게이션에 0월만 보였던 부분을 0000년 0월로 표시되도록 수정
- 네비게이션 이벤트 event/navigation.js 로 분리
