+++ client/src/App/index.js
@@ -8,6 +8,7 @@ import {
   HkbHistInputEvent,
   GraphMouseOver,
   GraphMouseOut,
+  NavigationEvent,
 } from './event';
 
 const pageRoute = (path) => {
@@ -15,66 +16,13 @@ const pageRoute = (path) => {
   layout.innerHTML = Route(path);
 };
 
-const activeNavigation = (target) => {
-  const navBtns = document.querySelectorAll('.nav_btn');
-  navBtns.forEach((navBtn) => navBtn.classList.remove('active'));
-  target.classList.add('active');
-};
-
-const setPageHistory = (target) => {
-  if (target.id === 'page_nav_hkb') {
-    window.history.pushState('hbk', '', '/');
-    setState('path', 'hbk');
-  } else if (target.id === 'page_nav_calendar') {
-    window.history.pushState('calendar', '', '/calendar');
-    setState('path', 'calendar');
-  } else if (target.id === 'page_nav_graph') {
-    window.history.pushState('graph', '', '/graph');
-    setState('path', 'graph');
-  }
-};
-
 const App = () => {
   const path = getState('path');
 
   const app = document.querySelector('.App');
 
-  // 액션 등록
-  app.addEventListener('click', (event) => {
-    const { target } = event;
-
-    // 페이지 네이게이션
-    if (target.classList.contains('nav_btn')) {
-      activeNavigation(target);
-      setPageHistory(target);
-    }
-
-    // navigation 월별 이동
-    // TODO : container 분리 후 navigation.js에 적용
-    const prevMonth = document.querySelector('.prev_month');
-    const nextMonth = document.querySelector('.next_month');
-
-    if (event.target === prevMonth) {
-      const currentMonth = getState('currentMonth');
-      let changeMonth = parseInt(currentMonth) - 1;
-      if (changeMonth < 1) {
-        changeMonth = 12;
-      }
-      setState('currentMonth', changeMonth);
-      const month = document.querySelector('.month');
-      month.textContent = `${changeMonth}월`;
-    }
-    if (event.target === nextMonth) {
-      const currentMonth = getState('currentMonth');
-      let changeMonth = parseInt(currentMonth) + 1;
-      if (changeMonth > 12) {
-        changeMonth = 1;
-      }
-      setState('currentMonth', changeMonth);
-      const month = document.querySelector('.month');
-      month.textContent = `${changeMonth}월`;
-    }
-  });
+  // 네비게이션 액션 등록
+  app.addEventListener('click', NavigationEvent);
 
   window.addEventListener('DOMContentLoaded', () => {
     if (window.location.pathname === '/graph') {
